arm64: time.c: fix the wrong format for printk
We will meet the compilor error for the current code:
  --------------------------------------------------
  time.c: In function ‘init_time’:
  time.c:131:12: error: format ‘%llx’ expects argument of
         type ‘long long unsigned int’, but argument 2 has
	 type ‘uint64_t {aka long unsigned int}’ [-Werror=format=]
      printk("Virtual Count register is %llx, freq = %d Hz\n", cntvct_at_init, counter_freq);
                   ^
  cc1: all warnings being treated as errors
  --------------------------------------------------

This patch fixes the compiling error.

+++ arch/arm/time.c
@@ -128,7 +128,8 @@ void init_time(void)
 
     __asm__ __volatile__("mrc p15, 0, %0, c14, c0, 0":"=r"(counter_freq));
     cntvct_at_init = read_virtual_count();
-    printk("Virtual Count register is %llx, freq = %d Hz\n", cntvct_at_init, counter_freq);
+    printk("Virtual Count register is %llx, freq = %d Hz\n",
+           (unsigned long long) cntvct_at_init, counter_freq);
 }
 
 void fini_time(void)
